Moved exception script to outer folder
