+++ startrak/exceptions.py
